+++ fluent_contents/models/db.py
@@ -6,6 +6,7 @@ from django.db import models
 from django.utils.translation import ugettext_lazy as _
 from polymorphic import PolymorphicModel
 from polymorphic.base import PolymorphicModelBase
+from fluent_contents import appsettings
 from fluent_contents.cache import get_rendering_cache_key
 from fluent_contents.models.managers import PlaceholderManager, ContentItemManager, get_parent_lookup_kwargs
 
@@ -65,7 +66,17 @@ class Placeholder(models.Model):
         Return the plugins which are supported in this placeholder.
         """
         from fluent_contents import extensions  # avoid circular import
-        return extensions.plugin_pool.get_plugins()
+
+        # See if there is a limit imposed.
+        slot_config = self.get_slot_config()
+        plugins = slot_config.get('plugins')
+        if not plugins:
+            return extensions.plugin_pool.get_plugins()
+        else:
+            try:
+                return extensions.plugin_pool.get_plugins_by_name(*plugins)
+            except extensions.PluginNotFound as e:
+                raise extensions.PluginNotFound(str(e) + " Update the plugin list of the FLUENT_CONTENTS_PLACEHOLDER_CONFIG['{0}'] setting.".format(self.slot))
 
 
     def get_content_items(self, parent=None):
@@ -79,6 +90,13 @@ class Placeholder(models.Model):
             item_qs = item_qs.filter(**get_parent_lookup_kwargs(parent))
 
         return item_qs
+
+
+    def get_slot_config(self):
+        """
+        Return the site-wide configuration associated with this slot.
+        """
+        return appsettings.FLUENT_CONTENTS_PLACEHOLDER_CONFIG.get(self.slot) or {}
 
 
     def get_absolute_url(self):