+++ fluent_contents/models/fields.py
@@ -3,6 +3,7 @@ from django.contrib.contenttypes.models import ContentType
 from django.db.models.query_utils import Q
 from django.utils.functional import lazy
 from django.utils.text import capfirst
+from fluent_contents import appsettings
 from fluent_contents.forms.fields import PlaceholderFormField
 from fluent_contents.models import Placeholder, ContentItem
 
@@ -143,9 +144,11 @@ class PlaceholderField(PlaceholderRelation):
         Initialize the placeholder field.
         """
         super(PlaceholderField, self).__init__(**kwargs)
-
         self.slot = slot
-        self._plugins = plugins
+
+        # See if a plugin configuration is defined in the settings
+        self._slot_config = appsettings.FLUENT_CONTENTS_PLACEHOLDER_CONFIG.get(slot) or {}
+        self._plugins = plugins or self._slot_config.get('plugins') or None
 
         # Overwrite some hardcoded defaults from the base class.
         self.editable = True
@@ -210,7 +213,7 @@ class PlaceholderField(PlaceholderRelation):
             try:
                 return extensions.plugin_pool.get_plugins_by_name(*self._plugins)
             except extensions.PluginNotFound as e:
-                raise extensions.PluginNotFound(str(e) + " Update the plugin list of '{0}.{1}'".format(self.model._meta.object_name, self.name))
+                raise extensions.PluginNotFound(str(e) + " Update the plugin list of '{0}.{1}' field or FLUENT_CONTENTS_PLACEHOLDER_CONFIG['{2}'] setting.".format(self.model._meta.object_name, self.name, self.slot))
 
 
     def value_from_object(self, obj):